And so begins aesalon . . .
Aesalon, once it is completed, will be a fully-functional
dynamically-allocated memory visualization tool. The basic design is
that it overloads malloc(), free(), etc. (using LD_PRELOAD) and passes
allocation data down a pipe to a recorder, which then logs and displays
the memory allocated in a graphical form of some kind. Since most
implementations of C++'s new operator simply call malloc(), this means
that Aesalon will work for both C and C++.

If I feel like playing around with the stack a little bit, it would also
be possible to determine the scope that the malloc() call occured from.
Of course, it will be in hexidecimal, but with random assorted
information from /proc/<pid>/ and read access to the executable, it is
not impossible to do.