Major problem in overloading malloc() . . .
. . . it seems dlopen()/dlsym()/etc. call malloc(). :P

Therefore, there needs to be a bootstrap shared library, that then calls
dlopen() with the required flags, RLTD_DEEPBIND | RLTD_GLOBAL. This means
that the overloaded malloc() will only be opened once all the pipe
initialization is completed.

A rewrite is, therefore, required. Aaah . . . how I love the smell of the
rm command in the evening.