Disabled writing to pipe unless environment variable is set.
Now, as for scope detection . . . the most obvious way, in my opinion,
is to check the return pointer of the malloc() function . . . hopefully,
that will be the address of the call instruction. Then, it is but a simple
matter to use nm to find the symbol that appears at the closest address
before that. Note that the scope resolution (and possibly demangling)
will be done in the main aesalon code, not inside the overload code.