The framework is now operating correctly.
In this case, the main aesalon executable is set to run a simple executable
located on my computer that simply allocates memory then frees it. If the
overload communication library is set to output to STDOUT rather than the pipe
passed in via AESALON_OVERLOAD_PIPE, then the malloc()/free() call information
is printed to stdout.