A bit of a problem with threads . . .
What appears to be happening, is when fork() is called, a new static instance
of EventQueue() (basically a Singleton instance) is created. So, this means
that anything pushed onto the queue in the one thread, the listening thread,
is not transferred to the instance in the other thread, the processing thread.

I'm not entirely certain how to fix this, though . . . besides, of course,
simply getting rid of the other thread . . .