Implemented MemoryBlock{,Manager}.
In other words, Aesalon now keeps track of the memory a program has allocated
using malloc()/free(). Not very exciting yet, but in the future it will help
greatly with the visualization.

And, speaking of visualization . . . I suppose it's time to start seriously
thinking about how to do it. What the plan in my head currently calls for is
simply a simple graphical interface that allows the user to display the total
memory allocated, the memory freed, the amount of allocations/frees per second
per scope, etc. However, in order to do that, a very complete GUI is required.

So, I suppose, in my rambling way, the point I'm trying to get across here is
"What graphical library should I use to display the memory data?". I could use
qt (probably version 4), and have a nice KDE-style interface, I could use GTK2
and have a GNOME-like interface . . . or I could use the SDL and write my own
GUI library . . .

Let's see, the advantages and disadvantages of each . . .

- GTK: looks ugly, but easy to use, but lacks some flexibility -- and is a C
    toolkit, not a C++ toolkit.
- qt: looks nicer than GTK, but more difficult to use. No lack of flexibility,
    is a C++ toolkit, but might be hard to render difficult graphics with.
- SDL: There are several options here, actually. I can write my own GUI, or I
    can use a pre-existing system. Advantages and disadvantages of each:
    - Custom GUI: difficult, would take a long time, but would end up with a
        custom-geared GUI that would fit the job.
    - Pre-existing GUI: easier, would take a shorter amount of time, but would
        have to work around the GUI quite a bit.

Overall, I'd say that using qt would probably be the best bet, since it is
possible to do most things that are required with it.

Alternatively, I could always do it the object-oriented way and support
multiple GUI systems . . . ;)