Fixed the listening thread, it terminates properly now.
This not only fixes the last memory leak, but it means that a zombie thread is
no longer left behind when aesalon runs . . . good news. ;)

Note that a Misc::UninitializedSingletonException is sometimes thrown on
execution, when the main process terminates while the listening thread is
pushing an event onto the EventQueue. Time to start using a mutex there, I
do believe.