It seems to me there is a way to solve the raging debate.
The obvious solution, really. Basically, what happens, is a command-line
argument is passed to the main executable, which is treated as the name of the
dynamic shared library to load for the main interface. A function would then
be called (extern "C"'d, of course, to remove mangling . . .), which
initializes the display library and parses the arguments. (argc/argv will of
course be passed to the initialization function . . .)

Then, the UI can initiate the aesalon interface. Clean, simple . . . well,
maybe not. In any case, it might work.