A basic, but flawed, beginning of an argument parser.
I'm actually not certain if I should be implementing an argument parser in
here, since the Qt libraries require arg{c,v} -- they parse information like
--display=remotehost:20, etc. However, an argument specifing the interface to
use has to be processed . . . aah, conundrums left and right.

Perhaps I should rethink the method of selecting the interface. The way I see
things, there's three ways I could do it . . .
- Hardcoding the display interface, and recompiling a separate executable for
    each. This has obvious drawbacks.
- Using an argument to pass a shared library name. This is the current design,
    and makes the most sense.
- Using an environment variable to store the path of a shared library. This,
    again, has advantages and disadvantages. In this case, the aesalon
    executable would be compiled to something like aesalon-main, then there
    would be bootstrap scripts to change the environment variable accordingly,
    like aesalon-qt, aesalon-ncurses, etc.

I'm beginning to think that I should just lose some flexibility in the current
design and make the argument parser slightly different . . . the problem is
this would mean that all display interfaces would have to take the same
arguments, and the Qt problem is still not fixed . . .

Oh yes, and since I believe I forgot to mention it; there are several display
interfaces on the back burner at the moment:
- Dummy interface, simply logs the allocation information to stdout/a file,
- ncurses interface, a fancy-shmancy text-based output, for use in non
    graphical environments,
- Qt interface, for those who cannot live without their graphics.

I think that the dummy interface would be a good one to implement first, aye?
It'll be the easiest, by far.

Heck with it. Inflexibility, here I come!