Re-implemented most of ArgumentParser.
Actually, removed the orignal Aesalon::ArgumentParser and moved it to
Aesalon::Misc::ArgumentParser, redesigned it, and did quite a bit more with
it. Note that a method of storing passed filenames (anything that's not an
argument or part of an argument) needs to be worked out. Probably just a
vector will suffice -- typedefed, of course.

Also added SmartPointer::to<ConvertedType>() and the matching
InvalidCastException to go with it. Quite useful, really.