Added FileArgument to ArgumentParser.
The argument parser should be working now. Best of all, it doesn't even
touch the original argc/argv. Note, though, that it's slightly inflexible; it
throws exceptions when it encounters unknown arguments. Perhaps it should be
permissive? Who knows.

Good news is, the dummy interface can now begin . . . with a slight change in
plans, of course. What is going to happen now is this. The different
interfaces will be built-in the executable; the argument simply selects which
member of a function pointer vector to call. This way, it can be selected at
buildtime what display interfaces to build. Complicated, but it should work.