Implemented the DisplayInitializer class.
Basically, the DisplayInitializer class takes the argument known by the
argument reference "display interface", gets its value, and creates an
instance of a derived type of a class that inherits from Singleton. As if that
wasn't enough, it then goes on to start the display interface. Where it goes
from there is up to the individual interface.

Note that, when an individual (e.g. me) wants to add a new interface to
aesalon, then they must add it in four separate places:
- DisplayInitializer.cpp, to include the header file for the new display
    interface,
- DisplayInitializer.h, to add it to the available_display_interfaces method,
- build.config, to enable its building,
- src/aesalon/display/CMakeLists.txt, for the aux_source_directory call and
    any additional libraries it needs.