I believe it is high time for a soliloquy about where aesalon is heading . . .
With the current design, aesalon will not be altogether useful, though it will
certainly be interesting. As is, it is nothing more than a glorified memory-
usage checker. However, if things are done properly, it should be perfectly
possible to turn it into something a little more useful.

Through extensive reading of various websites (valgrind's home page,
several Wikipedia web pages, many google searches, etc.), a few conversations,
and much thought, I have come to some conclusions.

In order to extend the usefulness of aesalon, a "tree view" of all the
currently dynamically allocated memory is the most obvious. In other words,
display the currently allocated memory, not in a flat form, but instead by
storage scope.

The most obvious way to do this is via a kernel-level reference counter, such
as valgrind. However, there is another, perhaps simpler, way.

gdb. Use gdb to set watchpoints on all the pointers in every scope. Then,
record the value of each pointer. Coupled with the current aesalon library,
this could turn into a fairly powerful tool -- view by block, view by scope.

Good enough for now. Time to get to it.