There are going to be a lot of redundancies in the codebase.
That is, once the gdb interface is implemented. Therefore, a lot of classes
need to be rewritten, modified, or removed.

A good portion of the Interface namespace needs to be rewritten, and a lot
more added to it. This is a good place to start. In fact, it is so badly
written (IMO), that it is time for a complete rewrite.

Hasta la vista, Interface.