There are a few problems that are cropping up right about now . . .
First of all, aesalon needs to be able to create a second thread. Come to
think of it, at least a third as well . . . I suppose I'll just use pthreads,
since there is the mutex support in the Misc::Singleton class already.

The second is the gdb parsing. I suppose that gdb --interpreter=mi will work
for the time being, but I need to verify exactly how that is going to be
parsed . . . perhaps a bison parser?

Oh yes, and I believe I may need a c++ parser to check what variables are
pointers . . . won't *that* be fun to implement.