Began implementing Aesalon::Interface::MemoryReference.
The most major problem there is how to store the scope information. Should it
be stored as a std::string? Or should it be stored as both a std::string and a
MemoryBlock, for the occurances where the block reference is stored within a
block? Even better, I suppose, would be a class called MemoryReferenceScope.
Aah, object-oriented programming.

Also fixed Misc::SmartPointer's operator-> and operator*; they were not const
operators, and therefore did not work for const SmartPointers. Bit of a
mistake on my part.

Fixed the .gitignore to have the proper executable, since everything was moved.
I cannot imagine how much larger the git repo must be now that there are a few
instances of the executable in it . . .

Also swapped BidirectionalPipe::is_open and ::is_connected().