A slight problem regarding gdb and malloc . . .
The problem being that I need to set a breakpoint on an overloaded symbol.
When I use "b malloc" in gdb, it sets the breakpoint on the version of malloc
in dl-minimal; I want the one in libc6. I'm sure there's an easy way, but I
suppose I'm not familliar enough with gdb to know.

And, to further complicate matters, Debian does not have the gdb documentation
on the DVD images, so I will have to wait a few days to get the documentation.
Or, I suppose I could install a virtual machine with Arch or some other distro
. . .

. . . or I could start working on the display module.

Which reminds me, I believe it may be a good idea to rename the "interface"
namespace and directory into "GDB", and rename "display" into "interface",
since the "display" is actually the user interface . . .

That would, of course, lead to confusion at first, at least for me, but I
think it makes more sense.