I've been musing on what interface could be renamed to . . .
The first one that sprang to mind was, of course, "GDB". While that does work,
there is a little more in there than just GDB . . . so, the next thing was
"platform". Platform does work, and I think slightly better than "interface".

Why not. Platform it is . . . search and replace, here I come . . .