Began framework for MemoryEvent, and changed EventQueue to be non-singleton.
Note the MemoryEvent::serialize(), as it will be used for logging purposes
later.

I believe I may have finally settled on a design that will work properly.
Basically, the primary aesalon executable simply is a dummy, command-line
program that executes an executable through gdb, and records memory
information. It can then either log the gathered data to disk, or it can
instead pass it along a network socket to the GUI executable.

Yes, a network socket. A network socket, in this case, works the best. The
primary reason that it works: aesalon can be used to monitor a program on
another system. In other words, remote monitoring of a program's memory
usage . . . not that it will really ever be used, but it is certainly
something to consider. Of course, if a network socket proves to be out of the
question, then a UNIX socket will have to suffice. How terrible.

This is the design used by many GUIs, having a command-line executable that
does the grunt work so the GUI can focus on doing GUI things, and not have to
worry about the actual work itself. I personally think it is a much more
elegant design than the current mess . . . but, then again, that could just be
my usual hallucinations . . .

We'll see.