So, this is what needs to be done in the future . . .
First of all, the buildsystem needs to be modified, which almost goes without
saying. The current design of src/interface, src/platform, src/misc needs to
be changed, quite dramatically, I think. Well, let's see what each new section
needs to do.

I think I will keep the design of src/interface and src/platform, come to
think of it -- they will simply create two separate executables. The target
created by src/interface will need to access some of the platform library,
though, which means that it needs to be separated out again, in order to avoid
as much redundant code as possible . . . or maybe not? I'm not entirely
certain -- design is most definitively not my forte.

The gist of it: two executables need to be created. One is the command-line
data-gatherer, the other is the GUI-based data-displayer. The GUI-based
data-displayer needs to be able to access some of the data structures and
classes that are contained within the data-gatherer, though, which is where
things are becoming complex. So, I suppose I need another source directory, as
aforementioned. src/interface is the qt4-based-GUI-data-displayer,
src/platform is the common source files, src/something is the command-line
data-gatherer, and src/misc is everything else that is common between the
three (SmartPointer, ReferenceCounter, etc.). Come to think of it, the
data-gatherer would actually be best named src/platform . . .

In any case, I am getting too tired to think clearly, and it is probably
showing through in my writing. Perhaps tomorrow I will be able to look on this
with a clear mind.