The results of further thought about splitting up aesalon again . . .
Basically, there is nothing that the command-line interface needs out of
Platform:: except BidirectionalPipe, and PipeListener. Hmm, come to think of
it, I believe PipeListener is about to become a pure virtual class sometime in
the nearby future. That way it can easily be overloaded as required.

In other words, the command-line executable needs only a small subset of
Platform::. This is, however, not a big issue, as the total size of
libaesalon_platform.a is only 1.9M (x86_64), and a stripped executable will
dramatically reduce the size of the library.

So, a new subdirectory needs to be created, one that creates the aesalon
executable. (Interface will create the aesalon-gui executable.) I'm not sure
what to call this subdirectory, though . . .

I'll think of something.