Solved the problem of what to call the new subdirectory.
Basically, the problem was solved by shifting where the GUI code will go. The
new directory is called gui, and the Aesalon::GUI namespace is where all of
the GUI-specific code will be placed. The directory src/interface, and the
corresponding namespace Aesalon::Interface will be the command-line data-
gatherer.