There is a slight problem cropping up, yet again.
Picture a typical execution of a program running under aesalon, begun from
within the GUI environment. The command-line gdb interface needs to send its
data along to the GUI, which it does so with a pipe. (It only needs to be one-
way, since the command-line program is doing all of the communication with
gdb.) Fine, so the pipe is created by the GUI, and the gdb interface simply
uses that pipe. But there is a catch . . . what happens when you want to
attach a GUI to an already-running gdb interface?

In other words, the gdb interface should create the pipe, not the GUI. Using a
named pipe, called something like "aesalon-<PID>". Then, when given the pid of
the gdb interface, it is possible to latch on to the pipe and get the data.

While this is attractive, it runs into issues with the other end . . . when
you start the gdb interface from within the GUI, the GUI has no idea what the
PID of the gdb interface is. Therefore, when the gdb interface is started with
a switch of some sort ("--gui-pipe-fd" perhaps), it simply sends the pid along
the pipe, then closes that pipe and creates the named pipe.

Or something like that.