Copied a very useful perl script, newclass.pl, from an old project of mine.
The script creates a new set of source files (.cpp, .h) in a given directory,
parsing the namespaces (takes Platform::Pipe and turns it into
src/platform/Pipe.{cpp,h}), and adds basic namespaces in (e.g. the
namespace Aesalon { namespace Platform {} } in the beginning of each source
file).

The original script was written by DWK, and I've simply modified it as
required.