I completely forgot about named-pipe blocking . . .
. . . which means that a slight redesign is in order, I believe. Basically,
what is happening is quite simple. From posix3 open, "An open() for writing-
only shall block the calling thread until a thread opens the file for
reading." I mean, there's even a dedicated errno entry: "ENXIO: O_NONBLOCK is
set, the named file is a FIFO, O_WRONLY is set, and no process has the file
open for reading." In other words, the gdb interface has to continually poll
the named pipe to see if there is anyone on the other end yet. Well, not
"continually", exactly, but fairly close.

Fine. If that's what it has to do, then that's what it has to do.