Added support for non-open pipes.
To be more exact, if the pipe is not open, on a write-pipe, and data is sent,
it will attempt to open the pipe. Failing that, it will just ignore the data.

Therefore, automagic communication . . . there is one difficulty, however. One
that, unfortunately, I just realized. The GUI will need all of the events that
have ever occured, not just the current ones. Therefore, a queue of messages
needs to be stored while nothing is reading on the other end of the pipe.

This is beginning to sound a little expensive.