Platform::PipeQueue is now implemented and used in the correct places.
The main problem, however, still remains. The main problem being the
communication between the GUI and the gdb interface. The pipe works well, but
only to a point. The GUI needs to know about every single event that has
occured, not just the ones from a specific point. In other words, the GUI
needs to be parsing the entire logfile that the gdb interface creates. Come to
think of it, that might actually be a good idea . . .

Well, maybe not. The current design functions well when there is only one GUI
interface monitoring the gdb interface. However, once there are multiple GUI
interfaces connected, and connecting half-way through, then the problems begin
to crop up, like now.

So, the way I am currently looking at this, there is only one way to solve
this problem, and that is to do a sort of hybrid between the two. Write to a
logfile, and also send communication data down a pipe of sorts . . . actually,
no, that wouldn't work, since multiple GUI clients could not connect. Hmm.

The problem that I see is this; the logfiles created by the gdb interface are
going to be absolutely huge, for a decently-sized application, and there is
going to be a large amount of I/O to those files; not a good idea for a file
on a hard drive.

So, there are two ways to solve this: either use a file to store everything,
and use the GUI to reconstruct the data afterwards, or only allow one
virtualization client to connect.

I'll think about it, I guess, and try to come up with a better solution.