After two days, I have still not come up with a better solution.
So: only one GUI client can connect at once. The gdb interface will log
everything to disk (using a queue to write out a certain number of events
at a time, to conserve disk access as much as possible), *only* when a logfile
is requested. This means that, when logging is not requested, it will simply
store all events in memory.

However, when a GUI client disconnects and reconnects (or a different one
connects), the gdb interface will *not* re-send all of the events to allow the
GUI interface to reconstruct the memory and references of the running program.
Instead, the gdb interface will only send the required events to reconstruct
the current memory references and allocated blocks, which should hopefully
save on "bandwidth" (amount of data transferred through the named pipe).

In other words, only one GUI client can connect to a single process. In order
to guard against multiple clients connecting, though . . . that might be a tad
tricky. The best solution I can think of involves using a second named pipe --
using the first one to get the name of the second (generated using tmpnam or a
hash of the date and time). However, that can wait for later, as it is not as
important. The most important thing is to just get the basics completed.

So, here goes nothing.