So, this is what is coming next . . .
. . . a full-blown parser. Hurrah.

Basically, there is a full definition of the gdb-mi2 interpreter output in the
info page. In order to fully utilize gdb's capabilities, I need to implement a
parser that will allow me to fully parse gdb's output.

So, a new subdirectory in src/interface/ needs to be created, one that is the
implementation of the gdb-mi2 parser (gdb-mi2 means gdb, Machine Interface
version 2), probably called "gdb" -- how imaginative.

The gdb parser will then be used by the gdb interface to create MemoryEvents
that are then, in turn, passed on to the GUI interface via the named pipe,
which are then displayed in a visual format. If that makes any sense.

Here goes nothing.