Having a bit of trouble with bisonc++.
bisonc++ *insists* on putting commas at the end of enumerator lists. How . . .
nice . . . of it to do so. Of course, with my compilation options, this is not
a warning, it's a serious offense, since I like to write compatible source
whenever possible. Or, at least I like to think I like to write compatible
source.

I'm not sure how to disable this warning -- there's nothing about it in the
gcc info page, nor the g++ info page. So, I suppose it's either manual
modification, complete disabling of pedantic warnings, or I work on something
else for the moment.

Of course, I could always just scrap the bisonc++ parser. The last option is
beginning to look more and more attractive, sad to say. I've never really
liked bison that much, but I do have to admit it gets the job done . . .