Fixed a memory leak or three.
Now, as for what happens next. What has been done so far is an excellent
start, but it is exactly that -- a start. Probably about a quarter of what is
needed has been done, to put things in perspective. So, what next?

The GUI. Enough said, as it is barely even begun.

The gdb interface pauses execution on watchpoints, but the actual watchpoints
themselves need to be set. Since this is on every single pointer, some method
of parsing variable definitions is required. The most obvious way to do this
is using a parser of some sort to parse variable declarations. However, this
might end up being more complex than it sounds -- this is not a simple
requirement, it's a full-blown research project, as a complete C++ parser
would need to be implemented . . . though I suppose I could try to throw
together something that might work in a pinch, it would be haphazard and
probably unstable.

Even the GUI alone is going to be at least another two thousand lines, let
alone a complete C++ parser . . . I suppose I might be able to utilize a pre-
existing parser of some sort, but I'm not sure how well that would work out.

I wonder if GDB has the capability to dump the contents of all local
variables? . . .