It truly is amazing what reading manuals can do . . .
There are two gdb commands that might just come in handy for aesalon. One is
'info locals', which displays all of the local variables in all the different
scopes on the stack, and the other is 'info variables', which displays _all
currently declared variables_, but takes about three minutes to compute.

While this does help in some ways, the fundamental problems are still
unsolved. Consider the following example:

class Container {
public:
    int *pointer;
};

int main(int argc, char *argv[]) {
    int variable = 42;
    Container c;
    c.pointer = &variable;
    return 0;
}

In the preceeding example, there are no local variables that are pointers, but
there is a pointer. So . . . how to solve this?

Well, come to think of it, there could be one (extremely unreliable) way to do
this. From the previous example, here is an example gdb session (extraneous
output removed):

(gdb) b main
Breakpoint 1 at 0x4006c7: file container_test.cpp, line 9.
(gdb) r
Starting program: /home/strange/c/container_test

Breakpoint 1, main (argc=1, argv=0x7fffffffe908) at container_test.cpp:9
9           int variable = 42;
(gdb) n
11          c.pointer = &variable;
(gdb) info local
variable = 42
c = {pointer = 0x7fffffffe900}
(gdb)

I'll extract the interesting information:

(gdb) info local
variable = 42
c = {pointer = 0x7fffffffe900}

Note that c, an instance of Container, shows the value of c.pointer, but it
does not specifically denote that it is a pointer. So, the question here is,
does GDB display the values of anything else as hexadecimal? If so, then this
path is denied. If not, then that could very well work. Simply place
breakpoints on functions (would need a parser for that as well, but a much
more basic one), and place watchpoints on variables that need them . . .

I'll think about it. There's got to be a more efficient way to do this.