Well, well. Watchpoints may not even be needed. Introducing: Variable Objects.
To quote gdb's info page, "Variable objects are [sic] "object-oriented" MI
interface for examining and changing values of expressions." Basically, this
is similar to watchpoints, but is (hopefully) easier to use, and additionally
"variable objects are specifically designed for simple and efficient
presentation in the frontend". In other words, seemingly custom-made.

The main reason I missed these before, I think, is because they are a MI-only
feature, and presently do not exist in the standard command-line interface (at
least, for the version of GDB I have, 7.0-1-debian).

So, how to utilize these, then. Unlike watchpoints, variable objects only
report their value has changed when upon request, which may complicate things.
Watchpoints were ideal because they reported that the value had changed
instantly. However, the major disadvantage of software watchpoints is their
speed -- inserting even one slows a program down by at least 200%.

Luckily, the MI2 GDB interpreter supports asynchronous commands. In other
words, commands can be executed while the program is executing. This is,
however, only supported on certain platforms. To quote the info page again,
"On some targets, GDB is capable of processing MI commands even while the
target is running." However, "many commands that access the target do not work
when the target is running." Therefore, a test may be required.

Either way -- watchpoints, the door is that way.