Did some experimentation with variable objects . . .
And the results are positive, so far. Basically, variable objects appear to be
watchpoints that only report if they've changed on demand, rather than halting
execution whenever they change, which is perfect. This means that, when
coupled with asyncronous execution (assuming that variable objects can be used
in async mode, see previous commit for details), the program's execution does
not need to be halted at all. Which means that on multi-core systems (like the
one I'm using now) the slowdown should be minimal (beyond gdb's, of course).

So, the BidirectionalPipe is going to need to be changed around, to using non-
blocking pipes. The aesalon executable will, every second or so, use GDB to
get the status of all the variables, and create MemoryEvents accordingly.

Enough writing. Not enough coding.