There is a fairly major problem . . .
If the current design goes through, then the vast majority of memory changes
will go unnoticed. That is, if the variable contents are only checked once per
second, then 99.something% of pointers will appear and disappear during that
period.

So, it's back to the drawing table for the same old issues.

Variable objects, I think, should still work. At least they are a better idea
than using 'info local' . . . The major issue here is: is speed or
throughness more important? Perfect coverage is possible with watchpoints (at
least theoretically), but watchpoints would slow down the executed program to
such a degree it would be horrendous on anything slower than a 3.0GHz
processor. Imperfect coverage is available in for form of variable objects,
but at a speed that is about as good as it could get.

Of course, it could be possible to only monitor certain variables using
variable objects; but that sort of defeats the entire purpose, doesn't it?
After all, the intent here is to display, in a graphical form, all of the
dynamically allocated memory of a program . . .

I suppose I could change the update delay on the variable objects to something
much smaller than one second -- more like 50 milliseconds perhaps -- but that
is, in essence, solving the wrong problem. The problem is not the delay, but
the fact that not every change will be reported . . .

So. What would happen if, per se, a simple C++ parser was implemented? One
just advanced enough to tell if an assignment changes a pointer? Then, the gdb
interface could set breakpoints on those lines, and all will work out?

Possible -- not likely, but still possible, as I'd prefer to avoid parsing any
actual source code, and try to abstract aesalon from the program being
executed as much as possible . . . which makes it easier to run on different
platforms, different programming languages, and even different "natural"
languages (languages with a spoken component).

I still think, for some deluded reason, that variable objects are the way to
go. However, I think I'll do some research online -- check if anyone else has
run into this problem, or something similar.