Created a Symbol class and apropriate nm parser.
This will be used, later, to get the addresses where memory is changed. See,
gdb provides a very nice command, 'x/i', which allows the user to display an
address as assembly. This can be used, before execution of the program begins
in gdb, to set breakpoints on appropriate lines. Through further use of
'info locals', it should be possible to determine which instructions affect
pointers, and which affect non-pointers.

Note the operative word, "should".