Fixed the StreamOutput factory.
Wonder of wonders, aesalon now runs, parses the output of nm and gdb
(partially for the latter, anyhow), and exits correctly . . . I can't actually
remember the last commit where it did that.

Now, the trick is to parse the GNU assembly that has been disassembled. That
will unfortunately be the real trick here.

Not to mention setting variable objects . . . but that's a problem for the
future.