Added code to actually set a breakpoint when changes_memory() returns true.
What has been done here, so far, is a very good start, I think. There are, of
course, many things left to do . . . but, as far as the reference counter
goes, there are two things left to do: actually parse the assembly, and
determine a way to see if a particular assignment or memory change involves a
pointer or not. Well, that and actually setting the variable objects on the
pointers . . .