Fixed a few minor bugs, here and there.
Aesalon now executes an executable, from start to finish, and doesn't halt
execution permanantly like it used to.

In other words, I think the SymbolParser and AssemblyParser classes are
finished for now; that means it's time to start on the hard stuff.

Namely, a way of figuring out what variables there are that are pointers.
Well, that and setting breakpoints on malloc(), free(), realloc(), and
calloc().