As mentioned before, there are problems with setting breakpoints on malloc.
Basically, rather than calling the malloc() in libc, programs call a local
symbol called "malloc@plt" (for those who are interested, objdump works well.)
malloc@plt, in turn, calls the malloc() function. However, for some reason gdb
is not detecting that malloc() is being called. This might be something to do
with the roundabout way that malloc() is being accessed:

0000000000400500 <malloc@plt>:
  400500:       ff 25 5a 06 20 00       jmp    QWORD PTR [rip+0x20065a]
    # 600b60 <_GLOBAL_OFFSET_TABLE_+0x20>
  400506:       68 01 00 00 00          push   0x1
  40050b:       e9 d0 ff ff ff          jmp    4004e0 <_init+0x18>

In other words, malloc() is never referenced directly, and this may be causing
gdb to trip up. While I suppose I could simply set the breakpoint at
malloc@plt (which, by the way, gdb cannot find), I do not know if this is
behaviour unique to x86_64, Linux, or even gcc. In short, to set a breakpoint
on malloc@plt may be unportable, something I wish to avoid if possible.

Unfortunately, I see no way around it.