Began an overload shared library again.
This seems like the most portable and simplest way of getting malloc
information . . .

The only problem is communicating with the aesalon library what the heck is
going on. I suppose a named pipe is in order -- *again*. That's two named
pipes per aesalon gdb interface instance. Fine, it'll have to do. I suppose I
should put the named pipes in /var/lib/aesalon or something, to stop clogging
up /tmp/ so much.

Note that the overload library requires a GNU extension to dlsym(). Also, I'll
probably think of a better way to do this, and I'll end up deleting the
overload library for a second time.