GDB is not behaving again.
Basically, the aesalon gdb interface is getting into an infinite loop waiting
for the results of '-interpreter-exec console "x/i <address>"' . . . as if
it's never receiving it, or as if the program is running. Neither is
happening, which makes it all the more strange.

On a side note, however, the overload library is working well, and writes all
of the correct information to the named pipe.