I have come to one conclusion: the gdb interface is a *mess*.
Therefore, it is high past time that I cleaned it up a little. That basically
means, as we all know, that a rewrite is in order. This time, at least, it's
not going to be a *complete* rewrite, but it's going to be pretty close. For
starters, Controller and Processor are the same thing -- there's no need to
have them separated out like they are.

This should hopefully clear up some confusion with the gdb debugging output.