Bit of a design issue here, regarding the GDB::Controller class.
Basically, each StringHandler (and derivatives) need to be able to access the
Controller instance on demand. Therefore, Controller.h needs to be included in
the StringHandler.h file. Therefore, it gets included in the StringManager
header . . . which is included by Controller.h . . .

It's a circular include list. Fun, and I don't know how to solve it . . .