A bit of a problem is cropping up with the observer-based system.
While it is a step forwards from the old system, all is not well. (yet . . .)
Basically, what is happening is this: the Controller sends requests for gdb to
output the assembly code at each symbol location. Fine. Now, the observer
comes along, and -- how does it know which symbol is being parsed? How can it
set the symbol to parsed, how can it recognize that the symbol has been
parsed?

I suppose one way is to re-run nm (in SymbolManager) without --demangle, and
from there get the scope name . . . but that would be a waste of memory, I
think. Or perhaps not . . . we'll see.