Changed the Aesalon program to return whatever the child returns.
I really should have thought of it before, but oh well. It's fixed now . . .

As for where aesalon is going next, that I do not know. There are many things
left to do, including the GUI, a method of setting variable objects on
pointers, several memory leaks to fix, and lots of documentation to write.

The memory leaks can wait; they are make-work, so to speak. The documentation
isn't urgent; I can hold 4000 lines in my head well enough on short-term. The
variable objects are going to be very annoying, as is the GUI. Which, I
suppose, leaves me with two options: the GUI and the variable objects. And the
GUI is sort of useless without data to display . . .